Add files via upload
They removed the implementation of Found logic to skip Droplets to be considered. I couldn't find a feasible solution so dropped the implementation. It still partitions segments into individual sets for general comparison. Adds a Key Value Error checker to ignore false detections outside the path. More explanation in the except description. Added a dynamic change to the trajectory for both curves and straights. The current issue is an Unbounded Error. Will work on that soon